ai-service/requirements.txt
fastapi>=0.115.4
uvicorn[standard]>=0.32.0
google-generativeai>=0.3.0
pydantic>=2.10.0
python-dotenv>=1.0.1
httpx>=0.27.2
groq>=0.11.0